PerformanceControl: add show power profile view
read-only for now
